// backend/src/enums/lead.enums.ts
export enum LeadSource {
  FACEBOOK_ADS = 'FACEBOOK_ADS',
  INSTAGRAM_ADS = 'INSTAGRAM_ADS',
  GOOGLE_ADS = 'GOOGLE_ADS',
  WEBSITE_FORM = 'WEBSITE_FORM',
  MANUAL_ENTRY = 'MANUAL_ENTRY',
  CSV_IMPORT = 'CSV_IMPORT',
}

export enum LeadStatus {
  NEW = 'NEW',
  AI_STARTED = 'AI_STARTED',
  AI_IN_PROGRESS = 'AI_IN_PROGRESS',
  QUALIFIED = 'QUALIFIED',
  HUMAN_APPROVAL_REQUIRED = 'HUMAN_APPROVAL_REQUIRED',
  SITE_VISIT_SCHEDULED = 'SITE_VISIT_SCHEDULED',
  CONVERTED = 'CONVERTED',
  LOST = 'LOST',
  ARCHIVED = 'ARCHIVED',
}

export enum TimelineEventType {
  LEAD_CREATED = 'LEAD_CREATED',
  WHATSAPP_STARTED = 'WHATSAPP_STARTED',
  CUSTOMER_REPLIED = 'CUSTOMER_REPLIED',
  AI_QUALIFIED = 'AI_QUALIFIED',
  HUMAN_ESCALATION = 'HUMAN_ESCALATION',
  REMINDER_SENT = 'REMINDER_SENT',
  SITE_VISIT_BOOKED = 'SITE_VISIT_BOOKED',
  LEAD_ASSIGNED = 'LEAD_ASSIGNED',
  STATUS_UPDATED = 'STATUS_UPDATED',
  NOTE_ADDED = 'NOTE_ADDED',
  CONVERTED = 'CONVERTED',
}